small fixes

+++ provider/bigquery.go
@@ -1234,7 +1234,7 @@ func (store *bqOfflineStore) GetTrainingSet(id ResourceID) (TrainingSetIterator,
 	return store.newbqTrainingSetIterator(iter), nil
 }
 
-func (store *bqOfflineStore) GetTrainingSetTestSplit(id ResourceID, testSize float64, shuffle bool, randomState int) (TrainingSetIterator, TrainingSetIterator, func() error, error) {
+func (store *bqOfflineStore) GetTrainingSetTestSplit(id ResourceID, testSize float32, shuffle bool, randomState int) (TrainingSetIterator, TrainingSetIterator, func() error, error) {
 	return nil, nil, nil, fmt.Errorf("not Implemented")
 }
