+++ provider/spark.go
@@ -2204,7 +2204,7 @@ func (spark *SparkOfflineStore) GetTrainingSet(id ResourceID) (TrainingSetIterat
 	return fileStoreGetTrainingSet(id, spark.Store, spark.Logger)
 }
 
-func (spark *SparkOfflineStore) GetTrainingSetTestSplit(id ResourceID, testSize float64, shuffle bool, randomState int) (TrainingSetIterator, TrainingSetIterator, func() error, error) {
+func (spark *SparkOfflineStore) GetTrainingSetTestSplit(id ResourceID, testSize float32, shuffle bool, randomState int) (TrainingSetIterator, TrainingSetIterator, func() error, error) {
 	return nil, nil, nil, nil
 }
