+++ provider/unit_test_store.go
@@ -287,6 +287,6 @@ func (m MockUnitTestOfflineStore) GetTrainingSet(id ResourceID) (TrainingSetIter
 	return nil, nil
 }
 
-func (m MockUnitTestOfflineStore) GetTrainingSetTestSplit(id ResourceID, testSize float64, shuffle bool, randomState int) (TrainingSetIterator, TrainingSetIterator, func() error, error) {
+func (m MockUnitTestOfflineStore) GetTrainingSetTestSplit(id ResourceID, testSize float32, shuffle bool, randomState int) (TrainingSetIterator, TrainingSetIterator, func() error, error) {
 	return nil, nil, nil, nil
 }
